+++ src/behaviors/Foreground/styles/ForegroundLayer.js
@@ -1,6 +1,13 @@
 import styled from 'styled-components';
+import { themeGet } from 'extensions';
 
 export default styled.div`
+  font-family: ${themeGet('fonts.default')};
+  color: ${themeGet('colors.text.default')};
+  font-size: ${themeGet('fontSizes.default')};
+  line-height: 1.5;
+  letter-spacing: 0.28px;
+
   position: absolute;
   top: 0;
   left: 0;