+++ src/components/BandwidthProvider/styles/GlobalStyle.js
@@ -1,7 +1,8 @@
-import { createGlobalStyle } from 'styled-components';
+import { createGlobalStyle, css } from 'styled-components';
 import datePickerCss from 'components/DatePicker/styles/baseCss';
+import { cssvars } from 'themes';
+import { themeGet } from 'extensions';
 
-/* eslint-disable no-unused-expressions */
 export default createGlobalStyle`
   * {
     box-sizing: border-box;
@@ -11,6 +12,12 @@ export default createGlobalStyle`
     width: 100%;
     min-height: 100vh;
     position: relative;
+    color: ${themeGet('colors.text.default')};
+    font-size: ${themeGet('fontSizes.default')};
+  }
+
+  :root {
+    ${props => cssvars(props.theme)}
   }
 
   ${datePickerCss}