+++ src/components/Select/index.js
@@ -1,1 +1,10 @@
-export { default } from './Select';
+import Select from './Select';
+import Searchable from './SearchableSelect';
+
+Select.Searchable = Searchable;
+
+export default Select;
+
+export {
+  default as popperMatchWidthModifier,
+} from './popperMatchWidthModifier';