+++ src/components/Select/popperMatchWidthModifier.js
@@ -1,8 +1,8 @@
 export default {
   fn: data => {
-    data.styles.width = data.offsets.popper.width = Math.round(
-      data.offsets.reference.width,
-    );
+    data.styles.width = data.offsets.popper.width =
+      data.offsets.reference.width;
+    // this +1 seems important, maybe due to borders?
     data.offsets.popper.left = data.offsets.reference.left;
     data.offsets.popper.right = data.offsets.reference.right;
     return data;