+++ src/components/Select/styles/Arrow.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import styled from 'styled-components';
-import { Icon } from 'components';
+import Icon from 'components/Icon';
 
 const Arrow = styled(Icon)`
   transform: ${props => (props.expanded ? 'rotate(180deg)' : 'rotate(0deg)')};