+++ src/components/Select/styles/ClearButton.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import styled, { keyframes } from 'styled-components';
-import { Icon } from 'components';
+import Icon from 'components/Icon';
 import { themeGet } from 'extensions';
 
 const fadeIn = keyframes`