+++ src/components/Select/styles/CurrentValue.js
@@ -1,11 +1,14 @@
 import React, { forwardRef } from 'react';
-import { Input, Loader } from 'components';
-import { Controls, Arrow, ClearButton } from './';
+import Input from 'components/Input';
+import Loader from 'components/Loader';
+import DefaultControls from './Controls';
+import DefaultArrow from './Arrow';
+import DefaultClearButton from './ClearButton';
 import styled from 'styled-components';
 import { themeGet } from 'extensions';
 
-const FakeInput = props => <Input as="button" {...props} />;
-const Wrapper = styled.div`
+const DefaultFakeInput = props => <Input as="button" {...props} />;
+const DefaultWrapper = styled.div`
   position: relative;
   cursor: pointer;
   height: 53px;
@@ -22,14 +25,14 @@ const Wrapper = styled.div`
     height: 100%;
   }
 `;
-const OverlayLoading = styled(Loader)`
+const DefaultOverlayLoading = styled(Loader)`
   position: absolute;
   left: 50% !important;
   top: 50% !important;
   transform: translate(-50%, -50%);
 `;
 
-export default forwardRef(
+const Unsearchable = forwardRef(
   (
     {
       isOpen,
@@ -41,6 +44,12 @@ export default forwardRef(
       loading,
       required,
       invalid,
+      Wrapper,
+      FakeInput,
+      OverlayLoading,
+      Controls,
+      ClearButton,
+      Arrow,
     },
     ref,
   ) => (
@@ -54,9 +63,31 @@ export default forwardRef(
         {loading ? <OverlayLoading /> : inputValue || placeholder}
       </FakeInput>
       <Controls>
-        {!!inputValue && !required && <ClearButton onClick={clearSelection} />}
+        {!!inputValue &&
+          !disabled &&
+          !required && <ClearButton onClick={clearSelection} />}
         <Arrow expanded={isOpen} />
       </Controls>
     </Wrapper>
   ),
 );
+
+Unsearchable.defaultProps = {
+  Wrapper: DefaultWrapper,
+  FakeInput: DefaultFakeInput,
+  Controls: DefaultControls,
+  OverlayLoading: DefaultOverlayLoading,
+  Arrow: DefaultArrow,
+  ClearButton: DefaultClearButton,
+};
+
+Unsearchable.styles = {
+  Wrapper: DefaultWrapper,
+  FakeInput: DefaultFakeInput,
+  Controls: DefaultControls,
+  OverlayLoading: DefaultOverlayLoading,
+  Arrow: DefaultArrow,
+  ClearButton: DefaultClearButton,
+};
+
+export default Unsearchable;