+++ stories/Select.js
@@ -5,6 +5,20 @@ import Select from 'components/Select';
 import Grid from 'layouts/Grid';
 
 const options = ['Apple', 'Banana', 'Lime', 'Lemon', 'Watermelon', 'Orange'];
+const complexOptions = [
+  {
+    label: 'One',
+    value: 1,
+  },
+  {
+    label: 'Two',
+    value: 2,
+  },
+  {
+    label: 'Three',
+    value: 3,
+  },
+];
 
 storiesOf('Select', module).add('types', () => (
   <Grid columns={1} maxSize="400px">
@@ -13,11 +27,17 @@ storiesOf('Select', module).add('types', () => (
     <h2>Disabled</h2>
     <Select disabled onChange={action('change')} options={options} />
     <h2>Searchable</h2>
-    <Select onChange={action('change')} searchable options={options} />
+    <Select.Searchable onChange={action('change')} options={options} />
     <h2>Required</h2>
     <Select onChange={action('change')} required options={options} />
     <h2>Loading</h2>
     <Select onChange={action('change')} loading options={options} />
+    <h2>Complex options</h2>
+    <Select
+      onChange={option => action('change')(option.value)}
+      options={complexOptions}
+      renderOption={option => option.label}
+    />
     <h2>Inside an overflow-hidden div</h2>
     <div style={{ overflow: 'hidden' }}>
       <Select onChange={action('change')} options={options} />